RF 00.02 | New Item > Create Pipeline Project > Create Pipeline Going From Manage Jenkins Page
Move testCreatePipelineGoingFromManageJenkinsPage() from NewItemTest to PipelineProjectTest
