+++ src/test/java/school/redrover/PipelineProjectTest.java
@@ -1,5 +1,6 @@
 package school.redrover;
 
+import org.apache.commons.lang3.RandomStringUtils;
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.ui.ExpectedConditions;
@@ -290,5 +291,25 @@ public class PipelineProjectTest extends BaseTest {
         String messageValue = resultMessage.getText();
 
         Assert.assertEquals(messageValue, newDescription);
+    }
+
+    @Test
+    public void testCreatePipelineGoingFromManageJenkinsPage() {
+        final String RANDOM_NAME_PROJECT = RandomStringUtils.randomAlphanumeric(8);
+
+        getDriver().findElement(By.linkText("Manage Jenkins")).click();
+        getDriver().findElement(By.linkText("New Item")).click();
+        getWait5().until(ExpectedConditions.visibilityOfElementLocated(By.xpath("//input[@name='name']"))).sendKeys(RANDOM_NAME_PROJECT);
+        getDriver().findElement(By.xpath("//span[text()='Pipeline']")).click();
+        getDriver().findElement(By.id("ok-button")).click();
+        getDriver().findElement(By.name("Submit")).click();
+        getDriver().findElement(By.xpath("//ol//a[@href='/']")).click();
+
+        List<String> jobList = getDriver().findElements(By.cssSelector(".jenkins-table__link"))
+                .stream()
+                .map(WebElement::getText)
+                .toList();
+
+        Assert.assertTrue(jobList.contains(RANDOM_NAME_PROJECT));
     }
 }
